+++ CHANGES.md
@@ -6,6 +6,9 @@
  - Visual scripting improvments in general + ai tool part.
  - MediaPipe implemented (hand model)
  - Test webRTC canvas capture to android TV  main instance for recceiver android-tv-cast.js/html
+ - Added First person shooter example (hang3d series - the-beast-hang3d)
+ - Base Position class changes, added `translateByXYZ`
+ - Micro optimisation : define CulledRenderPass only if culling activated from begin.
 
 [1.15.xx]
  - Added `npm install @google/generative-ai` like new ai tool provider.